pip-requirements.txt
torchvision==0.4.2
matplotlib==3.1.2
torch==1.3.1
numpy==1.17.4
scipy==1.4.1
tqdm==4.51.0
Pillow==8.0.1
PyYAML
